refactor(porygon): move interactors to common directory
