+++ packages/porygon/src/crud/modules/domain/ReadManyEntityInteractor.ts
@@ -1,5 +1,5 @@
+import { InteractorAsync } from '../../../common/modules/domain/InteractorAsync';
 import { FindEntityPort } from '../../port/application/FindEntityPort';
-import { InteractorAsync } from './InteractorAsync';
 
 export class ReadManyEntityInteractor<TModel, TQuery>
   implements InteractorAsync<TQuery, TModel[]>
