+++ packages/porygon/src/index.ts
@@ -1,7 +1,7 @@
+import { Interactor } from './common/modules/domain/Interactor';
+import { InteractorAsync } from './common/modules/domain/InteractorAsync';
 import { CrudModuleType } from './crud/models/domain/CrudModuleType';
 import { ModuleTypeToSymbolMap } from './crud/models/domain/ModuleTypeToSymbolMap';
-import { Interactor } from './crud/modules/domain/Interactor';
-import { InteractorAsync } from './crud/modules/domain/InteractorAsync';
 import { DomainCrudContainerModuleApi } from './crud/modules/iocuak/DomainCrudContainerModuleApi';
 import { CreateEntityPort } from './crud/port/application/CreateEntityPort';
 import { DeleteEntityPort } from './crud/port/application/DeleteEntityPort';
